refactor: move sanity types
