+++ components/Description.tsx
@@ -2,7 +2,7 @@ import Image from 'next/image'
 import { PortableText } from '@portabletext/react'
 import { urlForImage } from '@/sanity/lib/image'
 
-import type { Description as DescriptionProps } from '@/sanity/types/homepage'
+import type { Description as DescriptionProps } from '@/sanity/schemas/sections/description'
 
 export const Description = ({ content }: { content: DescriptionProps }) => {
   return (
