+++ components/Footer.tsx
@@ -2,7 +2,7 @@ import Link from 'next/link'
 import Image from 'next/image'
 
 import { urlForImage } from '@/sanity/lib/image'
-import type { Footer as FooterProps } from '@/sanity/types/homepage'
+import type { Footer as FooterProps } from '@/sanity/schemas/sections/footer'
 
 export const Footer = ({ content }: { content: FooterProps }) => {
   return (
