+++ components/MeetTheCharacter.tsx
@@ -2,7 +2,7 @@ import Image from 'next/image'
 import { PortableText } from '@portabletext/react'
 
 import { urlForImage } from '@/sanity/lib/image'
-import type { Character as CharacterProps } from '@/sanity/types/homepage'
+import type { Character as CharacterProps } from '@/sanity/schemas/sections/character'
 
 export const MeetTheCharacter = ({ content }: { content: CharacterProps }) => {
   return (
