+++ sanity/schemas/sections/trailer.ts
@@ -9,7 +9,11 @@ export const trailer = defineType({
       name: 'url',
       title: 'Url',
       type: 'url',
-      validation: (rule) => rule.required(),
+      validation: (rule) => rule.required().uri({ allowRelative: false }),
     }),
   ],
 })
+
+export type Trailer = {
+  url: string
+}
